+++ src/Unbound/Generics/LocallyNameless.hs
@@ -14,6 +14,7 @@ module Unbound.Generics.LocallyNameless (
   module Unbound.Generics.LocallyNameless.Bind,
   module Unbound.Generics.LocallyNameless.Embed,
   module Unbound.Generics.LocallyNameless.Rebind,
+  module Unbound.Generics.LocallyNameless.Rec,
   module Unbound.Generics.LocallyNameless.Fresh,
   module Unbound.Generics.LocallyNameless.LFresh,
   module Unbound.Generics.LocallyNameless.Subst
@@ -24,6 +25,7 @@ import Unbound.Generics.LocallyNameless.Name hiding (Bn, Fn)
 import Unbound.Generics.LocallyNameless.Bind hiding (B)
 import Unbound.Generics.LocallyNameless.Embed
 import Unbound.Generics.LocallyNameless.Rebind hiding (Rebnd)
+import Unbound.Generics.LocallyNameless.Rec
 import Unbound.Generics.LocallyNameless.Fresh
 import Unbound.Generics.LocallyNameless.LFresh
 import Unbound.Generics.LocallyNameless.Operations